refactor: UserAPI 분리(UserProfile, UserManagement)
