+++ src/test/java/com/playkuround/playkuroundserver/domain/user/api/UserManagementApiTest.java
@@ -38,7 +38,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @SpringBootTest
 @AutoConfigureMockMvc
 @ActiveProfiles("test")
-class UserApiTest {
+class UserManagementApiTest {
 
     @Autowired
     private ObjectMapper objectMapper;
